backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
python-dotenv>=1.0,<2.0
Pillow>=10.0,<11.0
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0
gunicorn>=21.0,<22.0
whitenoise>=6.6,<7.0
